# packages/api-client/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SceneProof API",
    "version": "0.1.0",
    "description": "Local development foundation. No authentication; do not expose publicly."
  },
  "paths": {
    "/api/v1/projects/{projectId}/findings/{findingId}/actions": {
      "parameters": [
        {
          "name": "projectId",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string",
            "format": "uuid"
          }
        },
        {
          "name": "findingId",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string",
            "format": "uuid"
          }
        }
      ],
      "post": {
        "operationId": "createFindingAction",
        "description": "An explicit INTENTIONAL_CHANGE authorizes one bounded targeted provider call. RESOLVE and DISMISS never call the provider. Scope must exactly match the original affected shots. Replaying the same requestId and payload returns the existing action without inference; changed payload conflicts. Failures retain the action and original finding. A new UUID explicitly authorizes a separate attempt.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateFindingAction"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Durable creator action and actual targeted result, if available",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FindingAction"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/Problem"
          }
        }
      },
      "get": {
        "operationId": "listFindingActions",
        "description": "Read-only immutable history, oldest first, at most 100 actions. supersedesActionId links the preceding effective judgement; only a successful result becomes effective. Original finding fields/evidence remain intact.",
        "responses": {
          "200": {
            "description": "Durable creator action and actual targeted result, if available",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FindingAction"
                  },
                  "minItems": 0,
                  "maxItems": 100
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/Problem"
          }
        }
      }
    },
    "/api/v1/projects/{projectId}/findings/{findingId}/actions/{actionId}": {
      "parameters": [
        {
          "name": "projectId",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string",
            "format": "uuid"
          }
        },
        {
          "name": "findingId",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string",
            "format": "uuid"
          }
        },
        {
          "name": "actionId",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string",
            "format": "uuid"
          }
        }
      ],
      "get": {
        "operationId": "getFindingAction",
        "responses": {
          "200": {
            "description": "Durable creator action and actual targeted result, if available",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FindingAction"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/Problem"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "CreateFindingAction": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "requestId",
          "type",
          "explanation",
          "scope",
          "affectedShotIds"
        ],
        "properties": {
          "requestId": {
            "type": "string",
            "format": "uuid"
          },
          "type": {
            "type": "string",
            "enum": [
              "INTENTIONAL_CHANGE",
              "RESOLVE",
              "DISMISS"
            ]
          },
          "explanation": {
            "type": "string",
            "minLength": 1,
            "maxLength": 2000
          },
          "scope": {
            "type": "string",
            "minLength": 1,
            "maxLength": 1000
          },
          "affectedShotIds": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "minItems": 1,
            "maxItems": 8
          }
        }
      },
      "TargetedResult": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "projectId",
          "originalFindingId",
          "schemaVersion",
          "outcome",
          "summary",
          "explanation",
          "evaluatedScope",
          "affectedShotIds",
          "evidenceFrameIds",
          "inspectedShots",
          "remainingIssue",
          "suggestedCorrection"
        ],
        "properties": {
          "projectId": {
            "type": "string",
            "format": "uuid"
          },
          "originalFindingId": {
            "type": "string",
            "format": "uuid"
          },
          "schemaVersion": {
            "type": "string",
            "enum": [
              "1"
            ]
          },
          "outcome": {
            "type": "string",
            "enum": [
              "INTENT_ACCEPTED",
              "ISSUE_REMAINS",
              "INSUFFICIENT_EVIDENCE"
            ]
          },
          "summary": {
            "type": "string",
            "minLength": 1,
            "maxLength": 1000
          },
          "explanation": {
            "type": "string",
            "minLength": 1,
            "maxLength": 3000
          },
          "evaluatedScope": {
            "type": "string",
            "minLength": 1,
            "maxLength": 1000
          },
          "affectedShotIds": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "minItems": 1,
            "maxItems": 8
          },
          "evidenceFrameIds": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "minItems": 1,
            "maxItems": 24
          },
          "inspectedShots": {
            "type": "array",
            "items": {
              "type": "object",
              "additionalProperties": false,
              "required": [
                "shotId",
                "frameIds"
              ],
              "properties": {
                "shotId": {
                  "type": "string",
                  "format": "uuid"
                },
                "frameIds": {
                  "type": "array",
                  "items": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "minItems": 1,
                  "maxItems": 3
                }
              }
            },
            "minItems": 1,
            "maxItems": 8
          },
          "remainingIssue": {
            "type": "string",
            "minLength": 0,
            "maxLength": 2000
          },
          "suggestedCorrection": {
            "type": "string",
            "minLength": 0,
            "maxLength": 3000
          }
        }
      },
      "FindingAction": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "id",
          "projectId",
          "findingId",
          "originalAnalysisRunId",
          "requestId",
          "type",
          "explanation",
          "scope",
          "affectedShotIds",
          "createdAt",
          "supersedesActionId",
          "reanalysis",
          "result"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "projectId": {
            "type": "string",
            "format": "uuid"
          },
          "findingId": {
            "type": "string",
            "format": "uuid"
          },
          "originalAnalysisRunId": {
            "type": "string",
            "format": "uuid"
          },
          "requestId": {
            "type": "string",
            "format": "uuid"
          },
          "type": {
            "type": "string",
            "enum": [
              "INTENTIONAL_CHANGE",
              "RESOLVE",
              "DISMISS"
            ]
          },
          "explanation": {
            "type": "string",
            "minLength": 1,
            "maxLength": 2000
          },
          "scope": {
            "type": "string",
            "minLength": 1,
            "maxLength": 1000
          },
          "affectedShotIds": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "minItems": 1,
            "maxItems": 8
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "supersedesActionId": {
            "type": [
              "string",
              "null"
            ],
            "format": "uuid"
          },
          "reanalysis": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AnalysisRun"
              },
              {
                "type": "null"
              }
            ]
          },
          "result": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TargetedResult"
              },
              {
                "type": "null"
              }
            ]
          }
        }
      },
...
"AnalysisUsage": {"type":"object","additionalProperties":false,"required":["inputTokens","outputTokens","totalTokens","cachedInputTokens","reasoningTokens","cacheWriteTokens"],"properties":{"inputTokens":{"type":["integer","null"],"minimum":0},"outputTokens":{"type":["integer","null"],"minimum":0},"totalTokens":{"type":["integer","null"],"minimum":0},"cachedInputTokens":{"type":["integer","null"],"minimum":0},"reasoningTokens":{"type":["integer","null"],"minimum":0},"cacheWriteTokens":{"type":["integer","null"],"minimum":0}}},
"AnalysisRun": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "projectId": {
            "type": "string",
            "format": "uuid"
          },
          "requestId": {
            "type": "string",
            "format": "uuid"
          },
          "status": {
            "type": "string",
            "enum": [
              "RUNNING",
              "SUCCEEDED",
              "FAILED"
            ]
          },
          "provider": {
            "type": "string",
            "enum": [
              "OPENAI"
            ]
          },
          "model": {
            "type": "string",
            "enum": [
              "gpt-6-astra"
            ]
          },
          "startedAt": {
            "type": "string",
            "format": "date-time"
          },
          "completedAt": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "failureCode": {
            "type": [
              "string",
              "null"
            ]
          },
          "failureMessage": {
            "type": [
              "string",
              "null"
            ]
          },
          "shotCount": {
            "type": "integer",
            "minimum": 0,
            "maximum": 8
          },
          "frameCount": {
            "type": "integer",
            "minimum": 0,
            "maximum": 24
          },
          "projectSummary": {
            "type": [
              "string",
              "null"
            ],
            "maxLength": 2000
          },
          "warnings": {
            "type": "array",
            "maxItems": 22,
            "items": {
              "type": "string",
              "maxLength": 500
            }
          },
          "usage": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AnalysisUsage"
              },
              {
                "type": "null"
              }
            ]
          },
          "providerResponseId": {
            "type": [
              "string",
              "null"
            ]
          },
          "providerRequestId": {
            "type": [
              "string",
              "null"
            ]
          },
          "kind": {
            "type": "string",
            "enum": [
              "SEQUENCE",
              "TARGETED"
            ]
          }
        },
        "required": [
          "id",
          "projectId",
          "requestId",
          "status",
          "provider",
          "model",
          "startedAt",
          "completedAt",
          "failureCode",
          "failureMessage",
          "shotCount",
          "frameCount",
          "projectSummary",
          "warnings",
          "usage",
          "providerResponseId",
          "providerRequestId",
          "kind"
        ]
      },
...
      "Problem": {
        "type": "object",
        "required": [
          "title",
          "status"
        ],
        "properties": {
          "type": {
            "type": "string",
            "format": "uri",
            "default": "about:blank",
            "description": "Omission means about:blank per RFC 9457."
          },
          "title": {
            "type": "string"
          },
          "status": {
            "type": "integer",
            "minimum": 400,
            "maximum": 599
          },
          "detail": {
            "type": "string"
          },
          "instance": {
            "type": "string"
          },
          "analysisRunId": {
            "type": "string",
            "format": "uuid",
            "description": "Durable run identifier when an admitted analysis fails."
          }
        }
      }
    },
    "responses": {
      "Problem": {
        "description": "RFC 9457 problem response",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        }
      }
    }
  }
}